Simulate an AZRmodel

To be updated! Many things could be added to this simulation. The simulation function expects the model to be augmented with R or C functions for model. If these are not present, the function will add them. If they are present, the function assumes that these are correct and match the model structure.

AZRsimulate(model, simtime = NULL, IC = NULL, parameters = NULL,
  paramnames = NULL, paramvalues = NULL, dosingTable = NULL,
  FLAGdosOut = FALSE, method = "lsode", atol = 1e-06, rtol = 1e-06)

Arguments

model

An AZRmodel

simtime

Simulation time vector. If scalar provided then 1001 simulation steps will be used.

IC

Vector with initial conditions. If not provided the initial conditions stored in the model will be used as defaults.

parameters

A named parameter vector to be used for simulation. Named parameters need to be present in the model.

paramnames

An alternative to providing a named vector with parameter values is to provide a vector with parameter names and a vector with corresponding parameter values (see next input argument)

paramvalues

A vector with parameter values is 'paramnames' is defined.

dosingTable

A dataframe defining a dosing table. Required columns: TIME, INPUT, DOSE, DURATION, LAGTIME

FLAGdosOut

If TRUE then dosing table information will be added to the output variable. If FALSE then it will not be added. This flag only has effect if a dosingTable is given as input argument.

method

String with the deSolve method for ODE integration (default: lsode as this is a solver that can reasonably deal with state constraints)

atol

Absolute tolerance for integration

rtol

Relative tolerance for integration

Value

AZRmodel with attached simulation functions (see above)

Examples

model <- AZRmodel(system.file("examples/NovakTyson.txt", package="AZRsim"))
x <- AZRsimulate(model,400)
x <- AZRsimulate(model,400,parameters=c(k1=0.5))
x <- AZRsimulate(model,0:400,IC=c(Cyclin=0.5))
x <- AZRsimulate(model,0:400,IC=c(Cyclin=0.5),parameters=c(k1=0.9))