Return evaluated RHS of AZRmodel ODEs

Calculates and returns the RHS of ODEs in an AZRmodel for given states, parameters, and time (needs to be scalar). Dosing tables not allowed. Useful to assess issues in a larger AZRmodel.

AZRxdotcalc(model, time = 0, states = NULL, parameters = NULL)

Arguments

model

An AZRmodel

time

Scalar time at which to evaluate the RHS

states

Names vector with state values at which to evaluate the ODE RHS Values for all states in the model need to be provided. If not provided then the initial conditions in the model will be used. If non-numeric then potential changes through parameter settings are taken into account.

parameters

A named parameter vector to be used for evaluation Named parameters need to be present in the model.

Value

Vector with evaluated RHS of ODEs

Examples

model <- AZRmodel(system.file("examples/NovakTyson.txt", package="AZRsim"))

#> Error in eval(expr, envir, enclos): could not find function "AZRmodel"

x <- AZRxdotcalc(model)

#> Error in methods::is(input, "azrmod"): object 'model' not found

x <- AZRxdotcalc(model,400)

#> Error in methods::is(input, "azrmod"): object 'model' not found

x <- AZRxdotcalc(model,400,parameters=c(k1=0.5))

#> Error in methods::is(input, "azrmod"): object 'model' not found