AZRmodel class

AZRmodel objects represent dynamic, ODE based models that can be simulated. The class constructor function AZRmodel() can create such objects from text file descriptions of these models. If no input argument is provided, an empty AZRmodel object is returned.

create_model(input = NULL, simFlag = TRUE)

Arguments

input

A string with the full filename (including absolute or relative path) to a text file, describing an AZRmodel.

simFlag

If set to TRUE (default) then the model will be prepared for simulations (dosing, constraints, etc.) and all required sim functions will be generated and added to the AZRmodel attributes. Choose this option when you want to simulate. In cases were the model structure and the information about inputs is important, then choose "FALSE".

Value

An object of class azr_model. If input is NULL, an empty object is returned.

Examples

create_model()

#> 	AZRmodel
#> 	========
#> 	Name:                       unnamed_model 
#> 	Number States:              0 
#> 	Number Parameters:          0 
#> 	Number Variables:           0 
#> 	Number Reactions:           0 
#> 	Number Functions:           0 

filename <- system.file(package="AZRsim","examples","NovakTyson.txt")
create_model(filename)

#> 	AZRmodel
#> 	========
#> 	Name:                       Novak-Tyson Model 
#> 	Number States:              9 
#> 	Number Parameters:          26 
#> 	Number Variables:           3 
#> 	Number Reactions:           19 
#> 	Number Functions:           0